requirements.txt
asgiref==3.11.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
python-decouple==3.8
sqlparse==0.5.4
drf-yasg==1.21.7
django-filter==24.2
setuptools
